--- request_normalize/__init__.py
@@ -1,2 +1,0 @@
-# ruff: noqa: F401
-from .request_normalize import normalize_body, normalize_headers, normalize_request, normalize_url
